Add example of unexpected behavior
When I try and make an entire test pending using `xdescribe`, code that
is put in any `before(:all)` blocks are still executed. This seems to
me like it should not happen since I'm saying that I want to skip that
whole set of tests. In my eyes as a user, I wouldn't expect anything in
that block to execute if it's being skipped.

I think this is a bug, but I'm not 100% sure. At the very least, it
seems surprising...